requirements.txt
discord.py>=2.0.0
python-dotenv
pymongo[srv]
gTTS
asyncio
seaborn
matplotlib
pytz
numpy
pandas
backoff
aiohttp
APScheduler
networkx # Added for relationship graph
PyNaCl # Needed for voice support
scipy # Added for kamada_kawai_layout in plotting